TMSOPTIMIZATION class moved to optim_tms
