+++ simnibs/simulation/optim_tms.py
@@ -1,3 +1,5 @@
+import glob
+import multiprocessing
 import os
 import copy
 import datetime
@@ -7,7 +9,15 @@ import pickle
 import itertools
 import pandas as pd
 import numpy as np
+import scipy.io
+
 import simnibs
+from ..msh import mesh_io
+from ..simulation import SESSION, SimuList, TMSLIST, save_matlab_sim_struct, FEMSystem, run_optim_sim
+from ..simulation.sim_struct import _set_up_global_solver, _finalize_global_solver
+from ..utils.file_finder import SubjectFiles
+from ..utils.matlab_read import try_to_read_matlab_field, remove_None
+from ..utils.simnibs_logger import logger
 
 from .. import msh
 from ..utils.simnibs_logger import logger
@@ -52,7 +62,7 @@ def eval_optim(simulations, target, tms_optim, res_fn, qoi="normE", elmtype='elm
     """
     # load tms_list if filename
     if type(tms_optim) == str:
-        tms_list_new = simnibs.simulation.TMSOPTIMIZATION()
+        tms_list_new = TMSOPTIMIZATION()
         tms_list_new.read_mat_struct(tms_optim)
         tms_optim = tms_list_new
 
@@ -84,7 +94,7 @@ def eval_optim(simulations, target, tms_optim, res_fn, qoi="normE", elmtype='elm
                                                                                    len(tms_optim.optimlist.pos))
         assert res.shape[-1] > 1, 'Only one simulation found. No optimization possible.'
 
-        if not tms_optim.target or tms_optim.target != target:
+        if not tms_optim.target or tms_optim.target != target or not tms_optim.target_idx:
             # read first mesh to get index of target element
             try:
                 # try to read mesh geom from .hdf5
@@ -478,12 +488,15 @@ def optimize_tms_coil_pos(tms_optim, target=None,
     # run simulations
     simnibs.utils.simnibs_logger.logger.info("Starting optimization: "
                                              "{} FEMs on {} cpus".format(len(tms_optim.optimlist.pos), n_cpu))
+
+    # create copy of tms_optim, as resuming changes tms_optim.optimlist.pos
+    tms_optim_org = copy.copy(tms_optim)
     simulations_fn = tms_optim.run(allow_multiple_runs=True, cpus=n_cpu)
 
     # evaluate all simulations
     tms_list_optim = eval_optim(simulations_fn,
                                 target,
-                                tms_optim,
+                                tms_optim_org,
                                 os.path.join(tms_optim.pathfem,
                                              tms_optim.optim_name + '_' + tms_optim.time_str + '.csv'),
                                 qoi=tms_optim.qoi)
@@ -498,3 +511,477 @@ def optimize_tms_coil_pos(tms_optim, target=None,
     return {'best_conds': tms_list_optim,
             'simulations': simulations_fn,
             'tms_optim': tms_optim}
+
+
+class TMSOPTIMIZATION(SESSION):
+    """This class defines the structure to optimize the TMS coil position/rotation for a given cortical target.
+
+    Use one TMSOPTIMIZATION per optimization target. That means:
+    - only one TMSLIST is accepted, with
+        - same mesh & anisotropy file for all positions
+        - same coil for all positions
+        - same intensity for all positions
+    """
+
+    def __init__(self, matlab_struct=None):
+        SESSION.__init__(self, matlab_struct)
+
+        # SESSION fields
+        del self.poslists
+        self.open_in_gmsh = False
+
+        # new fields
+        self.compress_hdf5 = None  # None | 'gzip'
+        self.cond = None
+        self.dAdt = None
+        self.didt = None
+        self.distance = None
+        self.fname_hdf5 = None
+        self.fnamecoil = None
+        self.mesh = None
+        self.n_sim = None
+        self.optim_name = ''
+        self.optimlist = None
+        self.qoi = 'E'
+        self.resume = False
+        self.save_fields = None  # 'normE' | 'E' | list of both
+        self.target = None
+        self.target_coil_matsim = None
+        self.target_idx = None
+        self.anisotropy_type = 'scalar'
+        self.write_mesh_geom = True  # add mesh geometry information to .hdf5
+
+        # target information
+        self.target = None
+        self.target_coil_matsim = None
+        self.angle_limits = None
+        self.handle_direction_ref = None
+        self.radius = None
+        self.resolution_angle = None
+        self.resolution_pos = None
+
+    def _set_logger(self, fname_prefix=None, summary=False):
+        """Set logfile name to self.optim_name"""
+        if not fname_prefix:
+            fname_prefix = self.optim_name
+        super()._set_logger(fname_prefix=fname_prefix)
+
+    def add_poslist(self, pl):
+        """ Adds a SimList object to the poslist variable
+
+        Parameters:
+        ----------------
+        pl: sim_struct.SimuList
+            SimuList object
+        """
+        if not isinstance(pl, SimuList):
+            raise TypeError('Elements in poslist must be subclasses of SimuList')
+
+        # we only want one poslist for the optimization
+        self.optimlist = pl
+        self._prepared = False
+
+    def read_mat_struct(self, mat):
+        if type(mat) == str:
+            mat = scipy.io.loadmat(mat)
+
+        super().read_mat_struct(mat)
+
+        self.anisotropy_type = try_to_read_matlab_field(mat, 'anisotropy_type', str, 'scalar')
+        self.cond = try_to_read_matlab_field(mat, 'cond', str, self.date)
+        self.compress_hdf5 = try_to_read_matlab_field(mat, 'compress_hdf5', str)
+        self.didt = try_to_read_matlab_field(mat, 'didt', float)
+        self.distance = try_to_read_matlab_field(mat, 'distance', float)
+        self.fnamecoil = try_to_read_matlab_field(mat, 'fnamecoil', str)
+        self.fname_hdf5 = try_to_read_matlab_field(mat, 'hdf5_fn', str)
+        if len(mat['optimlist']) > 0:
+            for PL in mat['optimlist'][0]:
+                if PL['type'][0] == 'TMSLIST':
+                    self.add_poslist(TMSLIST(PL))
+                else:
+                    raise IOError(
+                        "poslist type is not of type TMSLIST")
+        self.n_sim = try_to_read_matlab_field(mat, 'n_sim', int)
+        self.qoi = try_to_read_matlab_field(mat, 'qoi', str)
+
+        # target information
+        self.target = try_to_read_matlab_field(mat, 'target', np.array)
+        self.target_idx = try_to_read_matlab_field(mat, 'target_idx', int)
+        self.target_coil_matsim = try_to_read_matlab_field(mat, 'target_coil_matsim', np.array)
+        self.angle_limits = try_to_read_matlab_field(mat, 'angle_limits', list)
+        self.handle_direction_ref = try_to_read_matlab_field(mat, 'handle_direction_ref', list)
+        self.radius = try_to_read_matlab_field(mat, 'radius', float)
+        self.resolution_angle = try_to_read_matlab_field(mat, 'resolution_angle', float)
+        self.resolution_pos = try_to_read_matlab_field(mat, 'resolution_pos', float)
+
+        self.save_fields = try_to_read_matlab_field(mat, 'save_fields', list)  # 'normE', 'E', both
+        self.write_mesh_geom = try_to_read_matlab_field(mat, 'write_mesh_geom', bool)
+
+        if self.compress_hdf5 == '':
+            self.compress_hdf5 = None
+        self.optim_name = try_to_read_matlab_field(mat, 'optim_name', str, '')
+
+    def remove_poslist(self, number=None):
+        """Removes the specified poslist
+
+        Parameters:
+        ------------------------------
+        number: int
+        indice of postist to be removed
+        """
+
+        if number:
+            logger.warn("TMSOPTIMISATION.remove_poslist() does not eat number= argument.")
+        self.optimlist = None
+        self._prepared = False
+
+    def clear_poslist(self):
+        """ Removes all poslists
+        """
+        self.remove_poslist()
+        self._prepared = False
+
+    def sim_struct2mat(self):
+        """ Creates a dictionary to save a matlab structure with scipy.io.savemat()
+
+        Returns
+        --------------------
+        dict
+            Dictionaty for usage with scipy.io.savemat
+        """
+        mat = {}
+        mat['type'] = 'SESSION'
+        mat['date'] = remove_None(self.date)
+        mat['volfn'] = remove_None(self.volfn)
+        mat['subpath'] = remove_None(self.subpath)
+        mat['eeg_cap'] = remove_None(self.eeg_cap)
+        mat['fnamehead'] = remove_None(self.fnamehead)
+        mat['pathfem'] = remove_None(self.pathfem)
+        mat['fname_tensor'] = remove_None(self.fname_tensor)
+        mat['vol'] = self.vol.sim_struct2mat()
+        mat['map_to_vol'] = remove_None(self.map_to_vol)
+        mat['map_to_MNI'] = remove_None(self.map_to_MNI)
+        mat['map_to_fsavg'] = remove_None(self.map_to_fsavg)
+        mat['map_to_surf'] = remove_None(self.map_to_surf)
+        mat['fields'] = remove_None(self.fields)
+        mat['fiducials'] = self.fiducials.sim_struct2mat()
+
+        mat['poslist'] = []
+        # mat['cond'] = remove_None(self.cond)
+        mat['hdf5_fn'] = remove_None(self.fname_hdf5)
+        mat['didt'] = remove_None(self.didt)
+        mat['distance'] = remove_None(self.distance)
+        mat['fnamecoil'] = remove_None(self.fnamecoil)
+        mat['qoi'] = remove_None(self.qoi)
+        mat['save_fields'] = remove_None(self.save_fields)
+        mat['n_sim'] = remove_None(self.n_sim)
+        mat['write_mesh_geom'] = remove_None(self.write_mesh_geom)
+        mat['compress_hdf5'] = remove_None(self.compress_hdf5)
+        mat['anisotropy_type'] = remove_None(self.anisotropy_type)
+        mat['optim_name'] = remove_None(self.optim_name)
+
+        if self.optimlist:
+            mat['optimlist'] = self.optimlist.sim_struct2mat()
+
+        # target information
+        mat['angle_limits'] = remove_None(self.angle_limits)
+        mat['handle_direction_ref'] = remove_None(self.handle_direction_ref)
+        mat['radius'] = remove_None(self.radius)
+        mat['resolution_angle'] = remove_None(self.resolution_angle)
+        mat['resolution_pos'] = remove_None(self.resolution_pos)
+        mat['target'] = remove_None(self.target)
+        mat['target_coil_matsim'] = remove_None(self.target_coil_matsim)
+        mat['target_idx'] = remove_None(self.target_idx)
+
+        return mat
+
+    def _prepare(self):
+        """
+        Prepares session for optimizations.
+
+        - relative paths are made absolute,
+        - empty fields are set to default values,
+        - check if required fields exist
+        - check if TMSLIST parameters are set correctly (e.g. no different diDt within poslist, ...)
+        - non-changing fields from the single poslist are stored in self
+
+
+        """
+        if self._prepared:
+            logger.warn('TMSOPTIMIZATION already was prepared. Repeating all steps now.')
+            self._prepared = False
+        self.fnamehead = os.path.abspath(os.path.expanduser(self.fnamehead))
+        if not os.path.isfile(self.fnamehead):
+            raise IOError('Cannot locate head mesh file: %s' % self.fnamehead)
+
+        sub_files = SubjectFiles(self.fnamehead, self.subpath)
+        self.fnamehead = sub_files.fnamehead
+        self.subpath = sub_files.subpath
+
+        if not os.path.isdir(self.subpath):
+            logger.warning('Cannot locate subjects m2m folder')
+            logger.warning('some postprocessing options might fail')
+            self.subpath = None
+
+        if not self.fname_tensor:
+            self.fname_tensor = sub_files.tensor_file
+
+        self.fname_hdf5 = self.optim_name + '.hdf5'
+
+        if self.eeg_cap:
+            raise NotImplementedError('eeg_cap is not tested for TMS optimization.')
+
+        logger.info('Head Mesh: {0}'.format(self.fnamehead))
+        logger.info('Subject Path: {0}'.format(self.subpath))
+        self.pathfem = os.path.abspath(os.path.expanduser(self.pathfem))
+        logger.info('Simulation Folder: {0}'.format(self.pathfem))
+
+        assert self.optimlist, "No TMSLIST as optimlist provided."
+        if os.path.isfile(self.fnamehead):
+            mesh = mesh_io.read_msh(self.fnamehead)
+            mesh.fix_surface_labels()
+            self.optimlist.postprocess = self.fields
+            self.optimlist.fn_tensor_nifti = self.fname_tensor
+
+            self.optimlist._prepare()
+            if not self.optimlist.mesh:
+                self.optimlist.mesh = mesh
+
+            # do some checks
+            assert len(self.optimlist.pos)
+            assert np.all([pos.didt == self.optimlist.pos[0].didt for pos in self.optimlist.pos]), \
+                "For TMSOPTIMIZATION, use the same didt for all positions."
+
+            for pos in self.optimlist.pos:
+                if pos.name:
+                    pos.name = int(pos.name)
+            assert np.all([not pos.name or type(pos.name) == int for pos in self.optimlist.pos]), \
+                "Provide no positions names or integer vales"
+
+            # if no pos names are given, use index
+            if not self.optimlist.pos[0].name:
+                for i, pos in enumerate(self.optimlist.pos):
+                    pos.name = str(i)
+
+            assert np.all([pos.distance == self.optimlist.pos[0].distance for pos in self.optimlist.pos]), \
+                "For TMSOPTIMMIZATION, use the same coil distance for all positions."
+
+            # store some fields in self as they don't change over the optimization positions
+            self.didt = self.optimlist.pos[0].didt
+            self.distance = self.optimlist.pos[0].distance
+            self.mesh = self.optimlist.mesh
+            self.fnamecoil = self.optimlist.fnamecoil
+            self.n_sim = len(self.optimlist.pos)
+
+            # store fields from TMSOPTIMIZATION in TMSLIST
+            self.optimlist.anisotropy_type = self.anisotropy_type
+        else:
+            raise IOError(
+                'Could not find head mesh file: {0}'.format(self.fnamehead))
+
+        # hdf5 fn stuff
+        if not self.fname_hdf5:
+            self.fname_hdf5 = 'optim_results_{}.hdf5'.format(self.time_str)
+        if not os.path.isabs(self.fname_hdf5):
+            self.fname_hdf5 = os.path.join(self.pathfem, self.fname_hdf5)
+        if not self.fname_hdf5.endswith('.hdf5'):
+            self.fname_hdf5 += '.hdf5'
+
+        # build list of fields to save from self.fields value
+        if not self.save_fields:
+            if self.fields != 'eE':
+                logger.warn("Only 'eE' (that is normE and E) are tested for optimization.")
+            self.save_fields = [field for field in self.fields]
+
+            # replace e by normE
+            try:
+                idx = self.save_fields.index('e')
+                self.save_fields[idx] = 'normE'
+            except ValueError:
+                pass
+
+        # hdf5 file init
+        if os.path.exists(self.fname_hdf5) and not self.resume:
+            backup_fn = self.fname_hdf5[:-5] + '_' + self.time_str + '_backup' + '.hdf5'
+            logger.warn("{} already exists. Moving to {}.".format(self.fname_hdf5, backup_fn))
+            os.rename(self.fname_hdf5, backup_fn)
+        if self.write_mesh_geom and not self.resume:
+            self.mesh.write_hdf5(self.fname_hdf5, compression='gzip')
+
+        # resuming unfinished optimization tasks
+        if self.resume:
+            self.optimlist = self.get_resume_optimlist()
+        self._prepared = True
+
+    def get_resume_optimlist(self):
+        """
+        Build TMSLIST of unfinished simulations.
+
+        Returns
+        --------
+        TMSLIST
+        """
+        # get unfinished indices
+        unfinished = self.get_unfinished_sim_idx()
+        if not unfinished:
+            logger.warn("No unfinished simulations found. Start with resume=False to start new optimization.")
+        resume_tms_list = copy.copy(self.optimlist)
+        resume_tms_list.pos = []
+        for i in unfinished:
+            resume_tms_list.add_position(self.optimlist.pos[i])
+
+        return resume_tms_list
+
+    def get_unfinished_sim_idx(self):
+        """Checks self.hdf5_fn to find not processed simulations.
+
+        Returns
+        -------
+        set
+            Contains the .hdf5/data/field indices that are not processed.
+            May be empty.
+        """
+
+        paths = ['/nodedata/', 'elmdata/']
+        unfinished = set()
+        try:
+            with h5py.File(self.fname_hdf5, 'r') as f:
+                for field, path in itertools.product(self.save_fields, paths):
+                    try:
+                        g = f[path + field]
+                        unfinished = unfinished.union(np.where(g[-1, :, :] == 0)[1].tolist())
+                    except KeyError:
+                        pass
+        except OSError:
+            raise OSError("Cannot open {}. "
+                          "Start optimization with resume=False to start from scratch.".format(self.fname_hdf5))
+
+        if unfinished:
+            logger.info("Found {} (out of {}) unfinished simulations.".format(len(unfinished), self.n_sim))
+        return unfinished
+
+    def run(self, cpus=1, allow_multiple_runs=True, save_mat=True):
+        """ Run simulations in the current optimization.
+
+        Parameters
+        -----------
+        cpus: int (optional)
+            Number of cpus to use. Not necessarily will use all cpus. Default: 1
+        allow_multiple_runs: bool (optinal)
+            Whether to allow multiple runs in one folder. Default: False
+        save_mat: bool (optional)
+            Whether to save the .mat file of this structure
+
+        Returns
+        ---------
+        Writes the optimization results.
+
+        """
+        if not self.optim_name:
+            print("optim_name not set. Defaulting to 'optimization'")
+            self.optim_name = "optimization"
+
+        self._set_logger()
+        if not self._prepared:
+            self._prepare()
+        dir_name_sim = os.path.abspath(os.path.expanduser(self.pathfem))
+
+        if not self.optimlist.pos:
+            logger.info("No simulations left.")
+            return self.fname_hdf5
+
+        # some directory checks
+        if os.path.isdir(dir_name_sim):
+            g = glob.glob(os.path.join(dir_name_sim, '*optimization*.mat'))
+            if g and not allow_multiple_runs:
+                raise IOError('Found already existing simulation results in directory.'
+                              ' Please run the simulation in a new directory or delete'
+                              ' the simnibs_simulation*.mat files from the folder : {0}'.format(dir_name_sim))
+
+            # g = glob.glob(os.path.join(dir_name_sim, '*.hdf5'))
+            # if g and not allow_multiple_runs:
+            #     logger.warn('Found existing simulations in ({}*.hdf5).'.format(dir_name_sim))
+            logger.info('Saving optimization results in {0}'.format(dir_name_sim))
+        else:
+            logger.info('Running optimization results in new directory {0}'.format(dir_name_sim))
+            os.makedirs(dir_name_sim)
+
+        if save_mat:
+            save_matlab_sim_struct(self,
+                                   os.path.join(dir_name_sim,
+                                                self.optim_name + '_{}.mat'.format(self.time_str)))
+
+        self.cond2elmdata(force=True)
+
+        cpus = np.min((cpus, self.n_sim))
+
+        # set up solver
+        solver = FEMSystem.tms(self.mesh, self.cond)
+        solver.lock = multiprocessing.Lock()  # for the thread-save .hdf5 writes
+
+        if cpus > multiprocessing.cpu_count():
+            logger.warn(
+                "{} parallel jobs requested, but only {} cores found.".format(cpus, multiprocessing.cpu_count()))
+
+        def ec(e):
+            """Error callback for pool.apply_async()"""
+            raise e
+
+        starttime = datetime.datetime.now()
+        if cpus == 1:
+            _set_up_global_solver(solver)
+            # simulate each position
+            for pos in self.optimlist.pos:
+                run_optim_sim(pos,
+                              self.mesh,
+                              self.fnamecoil,
+                              self.didt,
+                              self.fname_hdf5,
+                              self.fields,
+                              self.cond,
+                              self.save_fields,
+                              self.n_sim,
+                              self.compress_hdf5)
+            _finalize_global_solver()
+
+        else:
+            with multiprocessing.Pool(processes=cpus,
+                                      initializer=_set_up_global_solver,
+                                      initargs=(solver,)) as pool:
+                sims = []
+                for pos in self.optimlist.pos:
+                    sims.append(pool.apply_async(run_optim_sim, (pos,
+                                                                 self.mesh,
+                                                                 self.fnamecoil,
+                                                                 self.didt,
+                                                                 self.fname_hdf5,
+                                                                 self.fields,
+                                                                 self.cond,
+                                                                 self.save_fields,
+                                                                 self.n_sim,
+                                                                 self.compress_hdf5), error_callback=ec))
+                pool.close()
+                pool.join()
+
+        duration = datetime.datetime.now() - starttime
+        logger.info('=====================================')
+        logger.info("SimNIBS finished running simulations: #{}, {} ({}/simulation)".format(self.n_sim, duration,
+                                                                                           duration / self.n_sim))
+
+        logger.info('Simulation results stored in {}'.format(self.fname_hdf5))
+        return self.fname_hdf5
+
+    def cond2elmdata(self, force=False):
+        """
+        Wrapper method for optimlist.cond2elmdata().
+
+        Parameters:
+        -----------
+        force: bool (false)
+            Set force to recompute conductivity information, even if already present.
+        """
+        assert isinstance(self.optimlist, TMSLIST), "No poslist present."
+        if not self.cond or force:
+            self.cond = self.optimlist.cond2elmdata()
+        else:
+            logger.info("Reusing already processed conductivity information.")
